fix: resolve trivial compiler warnings in core flavor
- Remove dead Elvis operator on non-null queue title (Items.kt)
- Remove redundant Int conversions (StringUtils.kt)
- Remove unnecessary safe calls on non-null receivers (Dialog.kt, HomeViewModel.kt, NewPipe.kt)
- Add explicit type argument to the details list to preserve nullable element typing (Dialog.kt)

+++ app/src/main/java/com/dd3boh/outertune/utils/StringUtils.kt
@@ -131,10 +131,10 @@ fun closestAlbumMatch(query: String, stringList: List<AlbumEntity>): AlbumEntity
 fun numberToAlpha(l: Long): String {
     val alphabetMap = ('A'..'J').toList()
     val weh = if (l < 0) "0" else l.toString()
-    val lengthStr = if (weh.length.toInt() < 10) {
-        "0" + weh.length.toInt()
+    val lengthStr = if (weh.length < 10) {
+        "0" + weh.length
     } else {
-        weh.length.toInt().toString()
+        weh.length.toString()
     }
 
     return (lengthStr + weh + "\u0000").map {
